ques-ans-chatbot/pyproject.toml
[project]
name = "ques-ans-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.201",
    "google-genai>=1.7.0",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
]
